Integrate MongoDB and refactor to-do App
Added the Spring Boot MongoDB starter dependency in the pom.xml and created a new ToDoRepository that extends MongoRepository. Refactored the ToDo entity class to work with MongoDB by removing id based operations and leveraging MongoDB's automatic id generation. Adjusted application properties to include MongoDB URI and Database name. Changed the deletion and updating of to-dos in the controller to use description instead of id.